refactor(functions): 🎨 Move change host functionality to be a util function
